backend/requirements.txt
aiohttp==3.8.5
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.3.0
yarl==1.9.2
frozenlist==1.4.0
aiosignal==1.3.1
anyio==4.0.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Automat==22.10.0
certifi==2023.7.22
cffi==1.15.1
constantly==15.1.0
cryptography==41.0.3
cssselect==1.2.0
defusedxml==0.7.1
dj-rest-auth==5.0.1
Django==4.2.4
django-allauth==0.56.1
django-cors-headers==4.2.0
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-extra-fields==3.7.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
feedparser==6.0.10
filelock==3.12.4
filetype==1.2.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==4.9.3
Markdown==3.4.4
multidict==6.0.4
oauthlib==3.2.2
openai==0.28.1
packaging==23.1
parsel==1.8.1
Pillow==10.0.1
Protego==0.3.0
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
Scrapy==2.11.0
service-identity==23.1.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
tldextract==3.6.0
tqdm==4.66.1
Twisted==22.10.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
openai==0.28.0
httpx~=0.25.0
feedparser~=6.0.10
scrapy==2.11.0
redis==5.0.1
celery==5.3.4
w3lib==2.1.2
zope.interface==6.0
